+++ CHANGELOG.md
@@ -5,6 +5,7 @@ Pearl changelog
 #### Added
 - B+ Tree indexes ([#84](https://github.com/qoollo/pearl/pull/93))
 - Range indexes ([#11](https://github.com/qoollo/pearl/pull/102))
+- Blob version validation ([#120](https://github.com/qoollo/pearl/pull/129))
 
 #### Changed
 - Rebuild corrupted index automatically during startup ([#94](https://github.com/qoollo/pearl/pull/94))