+++ CHANGELOG.md
@@ -1,20 +1,32 @@
 # Changelog
 Pearl changelog
 
+
 ## [Unreleased]
 #### Added
-- B+ Tree indexes ([#84](https://github.com/qoollo/pearl/pull/93))
-- Range indexes ([#11](https://github.com/qoollo/pearl/pull/102))
-- Move corrupted blobs into speacial directory ([#98](https://github.com/qoollo/pearl/pull/98))
-- Blob version validation ([#120](https://github.com/qoollo/pearl/pull/129))
+
 
 #### Changed
-- Rebuild corrupted index automatically during startup ([#94](https://github.com/qoollo/pearl/pull/94))
+
 
 #### Fixed
-- Create corrupted directory only when needed ([#94](https://github.com/qoollo/pearl/pull/131))
+
 
 #### Updated
+
+
+## [0.7.0] - 2021-10-02
+#### Added
+- B+ Tree indexes (#84)
+- Range indexes (#11)
+- Move corrupted blobs into speacial directory (#98)
+- Blob version validation (#120)
+
+#### Changed
+- Rebuild corrupted index automatically during startup (#94)
+
+#### Fixed
+- Create corrupted directory only when needed (#94)
 
 
 ## [0.6.2] - 2021-09-14